Added rudimentary SUPER SLOW Haskell client.
Network.Socket.connect is just ridonkulously slow when you're trying to
start a bunch of connections in parallel, because it does a spinloop
waiting for each connection to be created in sequence. Why is that? I
dunno, but I'll either have to switch to using some kind of nonblocking
I/O package or abandon my dreams of using Haskell for this.